+++ runzi/automation/scaling/toshi_api/scaled_inversion_solution.py
@@ -43,7 +43,7 @@ class ScaledInversionSolution(object):
 
     def _create_inversion_solution(self, filepath, produced_by, source_solution_id, mfd_table=None, meta=None, predecessors=None, metrics=None):
         qry = '''
-            mutation ($source_solution: ID!, $created: DateTime!, $digest: String!, $file_name: String!, $file_size: Int!, $produced_by: ID!, $predecessors: [PredecessorInput]) {
+            mutation ($source_solution: ID!, $created: DateTime!, $digest: String!, $file_name: String!, $file_size: BigInt!, $produced_by: ID!, $predecessors: [PredecessorInput]) {
               create_scaled_inversion_solution(input: {
                   source_solution: $source_solution
                   created: $created
